+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- project-model: allow expressions to be used in flow calls;
 - plugins: new built-in task `throw`;
 - concord-console: ability to filter Ansible hosts by their inventory
 group;